# backend/RoomHubB2B/openapi/roomhub-b2b.openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "RoomHub B2B API",
    "description" : "REST API for the RoomHub B2B commercial space marketplace.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Current RoomHub B2B server"
  } ],
...
  "paths" : {
    "/api/listings/{listingId}" : {
      "put" : {
        "tags" : [ "Listings" ],
        "summary" : "Update an owned commercial space listing",
        "operationId" : "updateListing",
        "parameters" : [ {
          "name" : "listingId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateListingRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Listing updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ListingResponseDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid listing request",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401" : {
            "description" : "Authentication required",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403" : {
            "description" : "Only landlords can manage listings",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "Listing not found",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Listings" ],
        "summary" : "Delete an owned commercial space listing",
        "operationId" : "deleteListing",
        "parameters" : [ {
          "name" : "listingId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Listing deleted"
          },
          "401" : {
            "description" : "Authentication required",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403" : {
            "description" : "Only landlords can manage listings",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "Listing not found",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409" : {
            "description" : "Listing has booking history",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/listings" : {
      "get" : {
        "tags" : [ "Listings" ],
        "summary" : "List published commercial spaces",
        "operationId" : "getListings",
        "responses" : {
          "200" : {
            "description" : "Published listings",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ListingResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Listings" ],
        "summary" : "Publish a commercial space listing",
        "operationId" : "publishListing",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateListingRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Listing published",
            "headers" : {
              "Location" : {
                "description" : "Published listing URI",
                "style" : "simple",
                "schema" : {
                  "type" : "string",
                  "format" : "uri"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ListingResponseDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid listing request",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401" : {
            "description" : "Authentication required",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403" : {
            "description" : "Only landlords can publish listings",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/listings/{listingId}/hide" : {
      "post" : {
        "tags" : [ "Listings" ],
        "summary" : "Hide an owned commercial space listing",
        "operationId" : "hideListing",
        "parameters" : [ {
          "name" : "listingId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Listing hidden"
          },
          "401" : {
            "description" : "Authentication required",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403" : {
            "description" : "Only landlords can manage listings",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "Listing not found",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/listings/{listingId}/activate" : {
      "post" : {
        "tags" : [ "Listings" ],
        "summary" : "Reactivate an owned commercial space listing",
        "operationId" : "activateListing",
        "parameters" : [ {
          "name" : "listingId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Listing activated"
          },
          "401" : {
            "description" : "Authentication required",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403" : {
            "description" : "Only landlords can manage listings",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "Listing not found",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UpdateListingRequestDto" : {
        "type" : "object",
        "description" : "Commercial space listing update request",
        "properties" : {
          "title" : {
            "type" : "string",
            "example" : "Updated meeting room",
            "maxLength" : 255,
            "minLength" : 0
          },
          "description" : {
            "type" : [ "string", "null" ],
            "example" : "Screen, flip chart and coffee area"
          },
          "city" : {
            "type" : "string",
            "example" : "Barnaul",
            "maxLength" : 100,
            "minLength" : 0
          },
          "address" : {
            "type" : "string",
            "example" : "Lenina Avenue, 10",
            "maxLength" : 255,
            "minLength" : 0
          },
          "pricePerHour" : {
            "type" : "number",
            "example" : 3000.0,
            "minimum" : 0.01
          },
          "capacity" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 24,
            "minimum" : 1
          },
          "spaceType" : {
            "type" : "string",
            "enum" : [ "MEETING_ROOM", "CONFERENCE_HALL", "CLASSROOM", "LOFT", "SHOWROOM" ],
            "example" : "MEETING_ROOM"
          },
          "imageUrl" : {
            "type" : [ "string", "null" ],
            "example" : "https://example.com/listing-42.jpg",
            "pattern" : "^https?://\\S+$"
          }
        },
        "required" : [ "address", "capacity", "city", "pricePerHour", "spaceType", "title" ]
      },
      "ListingResponseDto" : {
        "type" : "object",
        "description" : "Published commercial space listing",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Listing identifier",
            "example" : 42
          },
          "title" : {
            "type" : "string",
            "description" : "Listing title",
            "example" : "Meeting room in the city center"
          },
          "city" : {
            "type" : "string",
            "description" : "City where the space is located",
            "example" : "Barnaul"
          },
          "pricePerHour" : {
            "type" : "number",
            "description" : "Rental price per hour",
            "example" : 2500.0
          },
          "capacity" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Maximum number of guests",
            "example" : 20
          },
          "spaceType" : {
            "type" : "string",
            "description" : "Commercial space type",
            "enum" : [ "MEETING_ROOM", "CONFERENCE_HALL", "CLASSROOM", "LOFT", "SHOWROOM" ],
            "example" : "MEETING_ROOM"
          },
          "imageUrl" : {
            "type" : [ "string", "null" ],
            "description" : "Public listing image URL",
            "example" : "https://example.com/listing-42.jpg"
          },
          "description" : {
            "type" : [ "string", "null" ],
            "description" : "Detailed listing description"
          },
          "address" : {
            "type" : "string",
            "description" : "Street address",
            "example" : "Lenina Avenue, 10"
          },
          "ownerOrganizationId" : {
            "type" : [ "integer", "null" ],
            "format" : "int64",
            "description" : "Owner organization identifier",
            "example" : 17
          }
        }
      },
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "properties" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "CreateListingRequestDto" : {
        "type" : "object",
        "description" : "Commercial space listing publication request",
        "properties" : {
          "title" : {
            "type" : "string",
            "example" : "Meeting room in the city center",
            "maxLength" : 255,
            "minLength" : 0
          },
          "description" : {
            "type" : [ "string", "null" ],
            "example" : "Screen, flip chart and coffee area"
          },
          "city" : {
            "type" : "string",
            "example" : "Barnaul",
            "maxLength" : 100,
            "minLength" : 0
          },
          "address" : {
            "type" : "string",
            "example" : "Lenina Avenue, 10",
            "maxLength" : 255,
            "minLength" : 0
          },
          "pricePerHour" : {
            "type" : "number",
            "example" : 2500.0,
            "minimum" : 0.01
          },
          "capacity" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 20,
            "minimum" : 1
          },
          "spaceType" : {
            "type" : "string",
            "enum" : [ "MEETING_ROOM", "CONFERENCE_HALL", "CLASSROOM", "LOFT", "SHOWROOM" ],
            "example" : "MEETING_ROOM"
          },
          "imageUrl" : {
            "type" : [ "string", "null" ],
            "example" : "https://example.com/listing-42.jpg",
            "pattern" : "^https?://\\S+$"
          }
        },
        "required" : [ "address", "capacity", "city", "pricePerHour", "spaceType", "title" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}